Move extracted methods to new file
